Extract deployment-related funcs from controller.go to deployment.go
